backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
python-decouple==3.8
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core==0.1.10
pydantic==1.10.12
